requirements.txt
requests
camelot-py
pandas
opencv-python
matplotlib
seaborn
flask
beautifulsoup4